requirements.txt
openai-agents>=0.0.9
openai>=1.50.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0
arize-phoenix>=4.5.0
openinference-instrumentation-openai-agents>=0.1.0
openinference-instrumentation-openai>=0.1.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
aiofiles>=23.2.1
rich>=13.7.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
